refactor the code: break up drivers into different files
